UTOPIA-31 Admin microservice-JW

+++ src/main/java/com/ss/utopia/service/AirportService.java
@@ -18,35 +18,31 @@ public class AirportService{
 	@Autowired
 	private AirportRepository airportRepo;
 	
-	
-	public ResponseEntity<String> createAirport(String airportCode, String cityName) throws ClassNotFoundException, SQLException {
-		Collection<Airport> airportExist = airportRepo.findAirportByAirportCodeNative(airportCode);
+	public ResponseEntity<String> createAirport(Airport airport) throws ClassNotFoundException, SQLException {
+		Collection<Airport> airportExist = airportRepo.getAirportByIata(airport.getAirportCode());
 		if(!airportExist.isEmpty()) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("This airport already exists");
 		}
-		Airport airport = new Airport();
-		airport.setAirportCode(airportCode);
-		airport.setCityName(cityName);
 		airportRepo.save(airport);
 		return ResponseEntity.status(HttpStatus.ACCEPTED).body("Post Succeeded");
 	}
 	
 	public Iterable<Airport> getAirports() throws ClassNotFoundException, SQLException {		
-		Iterable<Airport> airports = airportRepo.findAll();
-		return airports;
+		return airportRepo.findAll();
 	}
 	
 	public ResponseEntity<Airport> getAirportById(String iata_id) {
-		Collection<Airport> airports = airportRepo.findAirportByAirportCodeNative(iata_id);
+		Collection<Airport> airports = airportRepo.getAirportByIata(iata_id.toUpperCase());
 		if(airports.isEmpty()) {
 			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
 		}
 		return ResponseEntity.status(HttpStatus.ACCEPTED).body(airports.iterator().next());
 	}
 	public ResponseEntity<String> updateAirport(String airportCode, String updateCityName) throws ClassNotFoundException, SQLException {
-		Collection<Airport> airportExist = airportRepo.findAirportByAirportCodeNative(airportCode);
+		Collection<Airport> airportExist = airportRepo.getAirportByIata(airportCode);
 		if(airportExist.isEmpty()) {
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Airport does not exist");		}
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Airport does not exist");		
+		}
 		Airport airport = new Airport();
 		airport.setAirportCode(airportCode);
 		airport.setCityName(updateCityName);
@@ -55,8 +51,13 @@ public class AirportService{
 
 	}
 	
-	public void deleteAirport(String iataId) throws ClassNotFoundException, SQLException {	
+	public ResponseEntity<String> deleteAirport(String iataId) throws ClassNotFoundException, SQLException {	
+		Collection<Airport> airportExist = airportRepo.getAirportByIata(iataId);
+		if(airportExist.isEmpty()) {
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Airport does not exist");		
+		}
 		airportRepo.deleteAirport(iataId);
+		return ResponseEntity.status(HttpStatus.ACCEPTED).body("Successfully deleted");
 	}
 }
